Fix: refactor implementation and design smells discovered in the core and structure modules
