+++ biojava-structure/src/main/java/org/biojava/nbio/structure/cluster/SubunitClusterer.java
@@ -71,14 +71,15 @@ public class SubunitClusterer {
 			for (int c1 = 0; c1 < clusters.size(); c1++) {
 				for (int c2 = clusters.size() - 1; c2 > c1; c2--) {
 					try {
+						SubunitClusterMerge subunitClusterMerge = new SubunitClusterMerge(clusters.get(c1));
 						if (params.isUseEntityIdForSeqIdentityDetermination() &&
-								clusters.get(c1).mergeIdenticalByEntityId(clusters.get(c2))) {
+								subunitClusterMerge.mergeIdenticalByEntityId(clusters.get(c2))) {
 							// This we will only do if the switch is for entity id comparison is on.
 							// In some cases it can save enormous amounts of time, e.g. for clustering full
 							// chains of deposited PDB entries. For instance for 6NHJ: with pure alignments it
 							// takes ~ 6 hours, with entity id comparisons it takes 2 minutes.
 							clusters.remove(c2);
-						} else if (clusters.get(c1).mergeSequence(clusters.get(c2), params)) {
+						} else if (subunitClusterMerge.mergeSequence(clusters.get(c2), params)) {
 							clusters.remove(c2);
 						}
 
@@ -96,7 +97,8 @@ public class SubunitClusterer {
 			for (int c1 = 0; c1 < clusters.size(); c1++) {
 				for (int c2 = clusters.size() - 1; c2 > c1; c2--) {
 					try {
-						if (clusters.get(c1).mergeStructure(clusters.get(c2), params)) {
+						SubunitClusterMerge subunitClusterMerge = new SubunitClusterMerge(clusters.get(c1));
+						if (subunitClusterMerge.mergeStructure(clusters.get(c2), params)) {
 							clusters.remove(c2);
 						}
 					} catch (StructureException e) {
@@ -122,7 +124,8 @@ public class SubunitClusterer {
 			for (int c1 = 0; c1 < clusters.size(); c1++) {
 				for (int c2 = clusters.size() - 1; c2 > c1; c2--) {
 					try {
-						if (clusters.get(c1).mergeStructure(clusters.get(c2), params))
+						SubunitClusterMerge subunitClusterMerge = new SubunitClusterMerge(clusters.get(c1));
+						if (subunitClusterMerge.mergeStructure(clusters.get(c2), params))
 							clusters.remove(c2);
 					} catch (StructureException e) {
 						logger.warn("Could not merge by Structure. {}",
